Phase 4 (B5/B6 + ns): unify DuplicateKeyPolicy, drop ColumnNumber, relocate internal helpers
B5 — DotEnv and INI had byte-identical duplicate-key enums
(LastWins/FirstWins/Disallowed). Replaced both with a single shared
Bodu.Text.DuplicateKeyPolicy. DelimitedDuplicateHeaderBehavior (adds
AllowDuplicates) and IniDuplicateSectionBehavior (merge semantics) are genuinely
different and kept as-is.

B6 — removed TextFormatException.ColumnNumber: it was only ever set to null
(line-oriented formats report LineNumber, Bencode reports Offset). Dropped the
property and the columnNumber ctor parameter; updated BencodeFormatException's
base call and the one test assertion.

ns — moved the formats-internal helpers TextThrowHelper and LineScanner out of
the now-public Bodu.Text namespace (which the new Bodu.Text library occupies)
into Bodu.Text.Formats, adding the using to their three consumers. The public
shared TextFormatException stays in Bodu.Text alongside DuplicateKeyPolicy.

B3 was assessed and dropped: the four *ParseOptions are readonly structs and
cannot share a base class.

Docs updated for all three. 837 regression / green.
