+++ Bodu.Text.Formats/src/Text/TextFormatException.cs
@@ -8,15 +8,15 @@ namespace Bodu.Text;
 
 /// <summary>
 /// Base class for parse-failure exceptions raised by the text-format parsers in <c>Bodu.Text.Formats</c>. Provides the
-/// common diagnostic surface — line, column, and offset — so callers can write a single handler that works across
-/// Bencode, Delimited, DotEnv, and INI sources.
+/// common diagnostic surface — line and offset — so callers can write a single handler that works across Bencode,
+/// Delimited, DotEnv, and INI sources.
 /// </summary>
 /// <remarks>
 /// <para>
 /// The location properties are advisory: each derived parser populates the fields it can identify at the point of
-/// failure. <see cref="LineNumber" /> uses <c>0</c> to mean "unknown" for backwards compatibility with the existing
-/// derived-type contracts, while <see cref="ColumnNumber" /> and <see cref="Offset" /> use <see langword="null" /> for
-/// the same purpose because they are new diagnostics.
+/// failure. <see cref="LineNumber" /> uses <c>0</c> to mean "unknown" while <see cref="Offset" /> uses
+/// <see langword="null" /> for the same purpose. Line-oriented formats report a line; byte-offset formats (Bencode)
+/// report an offset.
 /// </para>
 /// </remarks>
 public abstract class TextFormatException : FormatException
@@ -56,30 +56,16 @@ public abstract class TextFormatException : FormatException
     /// <param name="lineNumber">
     /// The 1-based line number on which the parse error was detected, or <c>0</c> when unknown.
     /// </param>
-    /// <param name="columnNumber">
-    /// The 1-based column number on which the parse error was detected, or <see langword="null" /> when unknown.
-    /// </param>
     /// <param name="offset">
     /// The 0-based offset from the start of the source at which the parse error was detected, or
     /// <see langword="null" /> when unknown.
     /// </param>
-    protected TextFormatException(string message, int lineNumber, int? columnNumber = null, int? offset = null)
+    protected TextFormatException(string message, int lineNumber, int? offset = null)
         : base(message)
     {
         LineNumber = lineNumber;
-        ColumnNumber = columnNumber;
         Offset = offset;
     }
-
-    /// <summary>
-    /// Gets the 1-based column number at which the parse error was detected, or <see langword="null" /> when the parser
-    /// cannot identify a specific column.
-    /// </summary>
-    /// <returns>
-    /// A positive integer identifying the source column, or <see langword="null" /> when the column is unknown or not
-    /// applicable to the format.
-    /// </returns>
-    public int? ColumnNumber { get; }
 
     /// <summary>
     /// Gets the 1-based line number at which the parse error was detected, or <c>0</c> when the error is not associated
